fix: autocompletion on multiuser homebrew
Aucompletion on ZSH is generated by the _zcompinit_ command,
which fails to run for security reasons if it finds executables
with group write permissions. Unfortunately, group write permissions
are a common way to make multiuser Homebrew setups work, so a multiuser
brew install may cause all users involved to lose autocompletion.
The fix is to run zcompinit with the -u (UNSAFE) argument.
This is a potential security risk if the users are untrusted.